Fix loop bounds in advect cell kernel launches
The issue is there are lots of macros which rely on xmin and ymin
to work out the array offset for memory access. However, these values
are passed in as the loop bounds, which are often -2, +2 in different
combinations. So, both the iteration counts were incorrect, *and* the
data being read!